+++ BaseLib/Algorithm.h
@@ -4,6 +4,6 @@
 
 
 void QSortStrings(const char** base, size_t count);
-int BinSearch(const char* search_str, const char** start, size_t count);
+intptr_t BinSearch(const char* search_str, const char** start, size_t count);
 
 #endif // _BASELIB_ALGORITHM_H_
